webapp/requirements-web.txt
Flask==3.0.2
requests==2.32.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
Werkzeug>=3.0.0

# OCR dependencies
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
numpy==1.26.4
